+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmObjectHelper.kt
@@ -195,15 +195,23 @@ internal object RealmObjectHelper {
             // The catch block should catch specific Core exceptions and rethrow them as Kotlin exceptions.
             // Core exceptions meaning might differ depending on the context, by rethrowing we can add some context related
             // info that might help users to understand the exception.
-        } catch (exception: RealmCorePropertyNotNullableException) {
-            throw IllegalArgumentException("Required property `${obj.className}.${obj.metadata[key]!!.name}` cannot be null")
-        } catch (exception: RealmCorePropertyTypeMismatchException) {
-            throw IllegalArgumentException("Property `${obj.className}.${obj.metadata[key]!!.name}` cannot be assigned with value '${value.value}' of wrong type")
-        } catch (exception: RealmCoreException) {
-            throw IllegalStateException(
-                "Cannot set `${obj.className}.$${obj.metadata[key]!!.name}` to `${value.value}`: changing Realm data can only be done on a live object from inside a write transaction. Frozen objects can be turned into live using the 'MutableRealm.findLatest(obj)' API.",
-                exception
-            )
+        } catch (exception: Throwable) {
+            throw CoreExceptionConverter.convertToPublicException(exception) { coreException: RealmCoreException ->
+                when (coreException) {
+                    is RealmCorePropertyNotNullableException -> {
+                        IllegalArgumentException("Required property `${obj.className}.${obj.metadata[key]!!.name}` cannot be null")
+                    }
+                    is RealmCorePropertyTypeMismatchException -> {
+                        IllegalArgumentException("Property `${obj.className}.${obj.metadata[key]!!.name}` cannot be assigned with value '${value.value}' of wrong type")
+                    }
+                    else -> {
+                        throw IllegalStateException(
+                            "Cannot set `${obj.className}.$${obj.metadata[key]!!.name}` to `${value.value}`: changing Realm data can only be done on a live object from inside a write transaction. Frozen objects can be turned into live using the 'MutableRealm.findLatest(obj)' API.",
+                            exception
+                        )
+                    }
+                }
+            }
         }
     }
 
@@ -212,7 +220,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         value: BaseRealmObject?,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         obj.checkValid()
@@ -224,7 +232,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         key: io.realm.kotlin.internal.interop.PropertyKey,
         value: BaseRealmObject?,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         setValueByKey(
@@ -238,7 +246,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         value: BaseRealmObject?,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         obj.checkValid()
@@ -250,7 +258,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         key: io.realm.kotlin.internal.interop.PropertyKey,
         value: BaseRealmObject?,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         if (value != null) {
@@ -267,7 +275,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         col: String,
         list: RealmList<Any?>,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         val existingList = getList<T>(obj, col)
@@ -472,7 +480,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         value: R,
-        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ALL,
         cache: ObjectCache = mutableMapOf()
     ) {
         obj.checkValid()
@@ -485,17 +493,15 @@ internal object RealmObjectHelper {
             CollectionType.RLM_COLLECTION_TYPE_NONE -> when (propertyMetadata.type) {
                 PropertyType.RLM_PROPERTY_TYPE_OBJECT -> {
                     if (obj.owner.schemaMetadata[propertyMetadata.linkTarget!!]!!.isEmbeddedRealmObject) {
-                        // FIXME Optimize make key variant of this
-                        setEmbeddedRealmObject(
+                        setEmbeddedRealmObjectByKey(
                             obj,
-                            propertyName,
+                            propertyMetadata.key,
                             value as BaseRealmObject?,
                             updatePolicy,
                             cache
                         )
                     } else {
-                        // FIXME Optimize make key variant of this
-                        setObject(obj, propertyName, value as BaseRealmObject?, updatePolicy, cache)
+                        setObjectByKey(obj, propertyMetadata.key, value as BaseRealmObject?, updatePolicy, cache)
                     }
                 }
                 else -> {