+++ test/base/src/commonMain/kotlin/io/realm/kotlin/entities/embedded/EmbeddedParentWithPrimaryKey.kt
@@ -22,11 +22,12 @@ import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.annotations.PrimaryKey
 
 // Convenience set of classes to ease inclusion of classes referenced by this top level model node
-val embeddedSchemaWithPrimaryKey = setOf(EmbeddedParentWithPrimaryKey::class, EmbeddedChild::class, EmbeddedInnerChild::class)
+val embeddedSchemaWithPrimaryKey = setOf(EmbeddedParentWithPrimaryKey::class, EmbeddedChildWithPrimaryKeyParent::class)
 
 class EmbeddedParentWithPrimaryKey : RealmObject {
     @PrimaryKey
-    var id: String? = null
-    var child: EmbeddedChild? = null
-    var childList: RealmList<EmbeddedChild> = realmListOf()
+    var id: Int? = null
+    var name: String? = "Realm"
+    var child: EmbeddedChildWithPrimaryKeyParent? = null
+    var children: RealmList<EmbeddedChildWithPrimaryKeyParent> = realmListOf()
 }